Moved mixin to new global mixin
